PowerShell script for Automation
SignAndPackage-Intune-App.ps1

This file utilizes whatever signature file you have stored for currentuser\my and signs your powershell scripts in your project directory prior to packaging the intune app. It's useful when you are working with a bunch of intune win app files simultaneously.

In its current stage you need to modify the paths at the top of the script to match where you are storing this tool and scripts.